Fixes auto-add from Gmail extraneous detail
Will now automatically strip the extra detail taken from the Google
gmail header. Does this based on there being a single gmail url in the
notes and the ASANA EMAIL ADDRESS APPEARING IN THE SUBJECT.

This means this won't work if your email address and your asana email
address are not the same. Shouldn't come up if you're doing Google
Login though.